Update Conditions documentation, Move Conditions outside of an offer so it can be reused
